pyproject.toml
[project]
name = "multi-agent-mlx"
version = "0.1.0"
description = "Local multi-agent software factory on Apple Silicon — zero API keys, 100% privacy"
requires-python = ">=3.12"
dependencies = [
    "agno",
    "openai",
    "mlx-lm>=0.25.2",
    "typer>=0.9.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
]

[project.scripts]
factory = "factory.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/factory"]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
